Add OSTreeRef and Arch field to Commit struct
We need to keep track of the ostree ref (such as rhel/x86_64/edge) which
will not change regardless of where the ostree remote is ultimately hosted.

Also we will need to keep track of the Arch to correlate a proper
candidate update image to a set of devices. This will also be useful for
ISO image creation.

Signed-off-by: Adam Miller <admiller@redhat.com>