update manager data model
If a device is not actively reachable or "online" via Cloud Connector, then
Playbook Dispatcher will 404 error when trying to schedule work for the device.
This behavior is the same for "offline" devices as well as invalid
device UUIDs, this means that we are going to have to manage the update
transaction. This change provides the basic data model for the update
transaction.

Signed-off-by: Adam Miller <admiller@redhat.com>