content-sources: API return HMS repositories
FIXES: THEEDGE-3158
To preserve compatibility: Return HMS repositories as thirdparty repos models with arbitrary/virtual ids, except for the ones already selected
we need to supply their real ids , this happens when we update an image, and we supply an imageID in url param.
The resolution is to have virtual ids superior to any id known in the current context.
Introduced a function GetImageByIDExtended that allow to have the image only with the data that are needed, and to not overcomplicate/overload this function for simple usage, when gormDB is not supplied (is nil) the core image is returned without any data queried.