content-sources: prohibit any changes to repos when feature flag is enabled.
FIXES: THEEDGE-3227
In the context of content-sources feature enabled, to avoid any confusion between local repos and content-sources one, block any modification to any local repo. for the API any repo must be modified only in content-sources.
In other words for the external world in this context we are using content-sources and have no feature in our application to modify or manage content-sources repos, this implies to return feature not available error when trying todo so on API.