images: Implement images storage cleanup.
The goal of the cleanup utility is to remove all unused data stored in AWS S3 bucket. we define unsed data by their image status:
- Image is soft-deleted, the cleanup will remove all stored data and delete the image forever.
- Image with status ERROR, the cleanup will remove only the stored data.
When removing the stored data the cleanup set the status of commit or repo or installer to STORAGE_CLEANED, this will enable us to identify the already cleaned stored data for images with status ERROR, and for deleted images when app is interrupted.

remain to implement the scheduler for this task.

FIXES: THEEDGE-3371