Add the ginkgolinter
The ginkgolinter is a linter that forces some coding standards and
prevents some issue when using ginkgo and gomega. Sew more datains in
https://github.com/nunnatsa/ginkgolinter.

In addition, some of the golangci-lint moved from the the github action
configuration to the .golangci.yml file, so it will be possible to run
the same set of linter locally, as in the github action.

And finally, replaced the `go-version` configuration in the github
action (the go setup step), with the `go-version-file: go.mod` so the go
setup will retreive the go version from the mod file, and will be
up-to-date when bumping golang version.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>