devicesview: return empty list when no devices
Return empty list in json [] when no devices where found.
The root cause where in the ReturnDevicesView, now we are creating and allocating to the devices slice (returnDevices), wich casue the devices slice to not be nil when no devices.
The end-point is returning the floowing payload when no data
{"count":0,"data":{"total":0,"devices":[],"enforce_edge_groups":false}}
FIXES: https://issues.redhat.com/browse/THEEDGE-3751