device-update: implement deviceUpdateImagesFilters
Implement deviceUpdateImagesFilters to allow to filter images for device update.
The issue raised when the user click on update for a device, on the device update view there is a filter that allow to filter update target images, the request was sent to backend, but as the backend feature was not implement, as the filter feature was not implemented, the same result was returned.
FIXES: https://issues.redhat.com/browse/THEEDGE-3795